string.py: adjust warning for isupper/islower
