+++ python/cudf/cudf/_libxx/cpp/strings/capitalize.pxd
@@ -1,5 +1,5 @@
 # Copyright (c) 2020, NVIDIA CORPORATION.
-from libcpp.memory cimpotr unique_ptr
+from libcpp.memory cimport unique_ptr
 
 from cudf._libxx.cpp.column.column cimport column
 from cudf._libxx.cpp.column.column_view cimport column_view
