+++ python/cudf/cudf/core/column/string.py
@@ -16,6 +16,15 @@ import cudf._libxx as libcudfxx
 import cudf._libxx.string_casting as str_cast
 from cudf._lib.nvtx import nvtx_range_pop, nvtx_range_push
 from cudf._libxx.null_mask import bitmask_allocation_size_bytes
+from cudf._libxx.strings.capitalize import (
+    capitalize as cpp_capitalize,
+    title as cpp_title,
+)
+from cudf._libxx.strings.case import (
+    swapcase as cpp_swapcase,
+    to_lower as cpp_to_lower,
+    to_upper as cpp_to_upper,
+)
 from cudf._libxx.strings.char_types import (
     is_alnum as cpp_is_alnum,
     is_alpha as cpp_is_alpha,
@@ -431,20 +440,6 @@ class StringMethods(object):
             **kwargs,
         )
 
-    def lower(self, **kwargs):
-        """
-        Convert strings in the Series/Index to lowercase.
-
-        Returns
-        -------
-        Series/Index of str dtype
-            A copy of the object with all strings converted to lowercase.
-        """
-
-        return self._return_or_inplace(
-            self._column.nvstrings.lower(), **kwargs
-        )
-
     # def slice(self, start=None, stop=None, step=None, **kwargs):
     #     """
     #     Returns a substring of each string.
@@ -550,8 +545,8 @@ class StringMethods(object):
         when operating on data which contains non-alphabetical characters.
         """
         warnings.warn(
-            "StringColumn method `isupper` does not conform to standard"
-            + "python string logic"
+            "isupper currently returns False for non-cased characters whereas"
+            + "Pandas returns True, this will be fixed in the near future"
         )
 
         return self._return_or_inplace(cpp_is_upper(self._column))
@@ -571,11 +566,87 @@ class StringMethods(object):
         when operating on data which contains non-alphabetical characters.
         """
         warnings.warn(
-            "StringColumn method `islower` does not conform to standard "
-            + "python string logic"
+            "islower currently returns False for non-cased characters whereas"
+            + "Pandas returns True, this will be fixed in the near future"
         )
 
         return self._return_or_inplace(cpp_is_lower(self._column))
+
+    def lower(self):
+        """
+        Convert each string to lowercase.
+        This only applies to ASCII characters at this time.
+
+        Examples
+        --------
+        >>> import nvstrings
+        >>> s = nvstrings.to_device(["Hello, Friend","Goodbye, Friend"])
+        >>> print(s.lower())
+        ['hello, friend', 'goodbye, friend']
+
+        """
+        return self._return_or_inplace(cpp_to_lower(self._column))
+
+    def upper(self):
+        """
+        Convert each string to uppercase.
+        This only applies to ASCII characters at this time.
+
+        Examples
+        --------
+        >>> import nvstrings
+        >>> s = nvstrings.to_device(["Hello, friend","Goodbye, friend"])
+        >>> print(s.upper())
+        ['HELLO, FRIEND', 'GOODBYE, FRIEND']
+
+        """
+        return self._return_or_inplace(cpp_to_upper(self._column))
+
+    def capitalize(self):
+        """
+        Capitalize first character of each string.
+        This only applies to ASCII characters at this time.
+
+        Examples
+        --------
+        >>> import nvstrings
+        >>> s = nvstrings.to_device(["hello, friend","goodbye, friend"])
+        >>> print(s.capitalize())
+        ['Hello, friend", "Goodbye, friend"]
+
+        """
+        return self._return_or_inplace(cpp_capitalize(self._column))
+
+    def swapcase(self):
+        """
+        Change each lowercase character to uppercase and vice versa.
+        This only applies to ASCII characters at this time.
+
+        Examples
+        --------
+        >>> import nvstrings
+        >>> s = nvstrings.to_device(["Hello, Friend","Goodbye, Friend"])
+        >>> print(s.swapcase())
+        ['hELLO, fRIEND', 'gOODBYE, fRIEND']
+
+        """
+        return self._return_or_inplace(cpp_swapcase(self._column))
+
+    def title(self):
+        """
+        Uppercase the first letter of each letter after a space
+        and lowercase the rest.
+        This only applies to ASCII characters at this time.
+
+        Examples
+        --------
+        >>> import nvstrings
+        >>> s = nvstrings.to_device(["Hello friend","goodnight moon"])
+        >>> print(s.title())
+        ['Hello Friend', 'Goodnight Moon']
+
+        """
+        return self._return_or_inplace(cpp_title(self._column))
 
     def slice_from(self, starts=0, stops=0, **kwargs):
         """
@@ -1167,21 +1238,17 @@ class StringColumn(column.ColumnBase):
         # the sum method for Numeric Series
         return self.nvstrings.join().to_host()[0]
 
-    def upper(self):
-        return self._return_or_inplace(libcudfxx.case.to_upper(self._column))
-
-    def lower(self):
-        return self._return_or_inplace(libcudfxx.case.to_lower(self._column))
-
     def capitalize(self):
-        return self._return_or_inplace(
-            libcudfxx.capitalize.capitalize(self._column)
-        )
+        return None
+        # return self._return_or_inplace(
+        #     libcudfxx.capitalize.capitalize(self._column)
+        # )
 
     def title(self):
-        return self._return_or_inplace(
-            libcudfxx.capitalize.title(self._column)
-        )
+        return None
+        # return self._return_or_inplace(
+        #     libcudfxx.capitalize.title(self._column)
+        # )
 
     @property
     def is_unique(self):
