+++ python/nvstrings/nvstrings.py
@@ -1648,97 +1648,6 @@ class nvstrings:
             rtn = nvstrings(rtn)
         return rtn
 
-    def lower(self):
-        """
-        Convert each string to lowercase.
-        This only applies to ASCII characters at this time.
-
-        Examples
-        --------
-        >>> import nvstrings
-        >>> s = nvstrings.to_device(["Hello, Friend","Goodbye, Friend"])
-        >>> print(s.lower())
-        ['hello, friend', 'goodbye, friend']
-
-        """
-        rtn = pyniNVStrings.n_lower(self.m_cptr)
-        if rtn is not None:
-            rtn = nvstrings(rtn)
-        return rtn
-
-    def upper(self):
-        """
-        Convert each string to uppercase.
-        This only applies to ASCII characters at this time.
-
-        Examples
-        --------
-        >>> import nvstrings
-        >>> s = nvstrings.to_device(["Hello, friend","Goodbye, friend"])
-        >>> print(s.upper())
-        ['HELLO, FRIEND', 'GOODBYE, FRIEND']
-
-        """
-        rtn = pyniNVStrings.n_upper(self.m_cptr)
-        if rtn is not None:
-            rtn = nvstrings(rtn)
-        return rtn
-
-    def capitalize(self):
-        """
-        Capitalize first character of each string.
-        This only applies to ASCII characters at this time.
-
-        Examples
-        --------
-        >>> import nvstrings
-        >>> s = nvstrings.to_device(["hello, friend","goodbye, friend"])
-        >>> print(s.capitalize())
-        ['Hello, friend", "Goodbye, friend"]
-
-        """
-        rtn = pyniNVStrings.n_capitalize(self.m_cptr)
-        if rtn is not None:
-            rtn = nvstrings(rtn)
-        return rtn
-
-    def swapcase(self):
-        """
-        Change each lowercase character to uppercase and vice versa.
-        This only applies to ASCII characters at this time.
-
-        Examples
-        --------
-        >>> import nvstrings
-        >>> s = nvstrings.to_device(["Hello, Friend","Goodbye, Friend"])
-        >>> print(s.swapcase())
-        ['hELLO, fRIEND', 'gOODBYE, fRIEND']
-
-        """
-        rtn = pyniNVStrings.n_swapcase(self.m_cptr)
-        if rtn is not None:
-            rtn = nvstrings(rtn)
-        return rtn
-
-    def title(self):
-        """
-        Uppercase the first letter of each letter after a space
-        and lowercase the rest.
-        This only applies to ASCII characters at this time.
-
-        Examples
-        --------
-        >>> import nvstrings
-        >>> s = nvstrings.to_device(["Hello friend","goodnight moon"])
-        >>> print(s.title())
-        ['Hello Friend', 'Goodnight Moon']
-
-        """
-        rtn = pyniNVStrings.n_title(self.m_cptr)
-        if rtn is not None:
-            rtn = nvstrings(rtn)
-        return rtn
-
     def index(self, sub, start=0, end=None, devptr=0):
         """
         Same as find but throws an error if arg is not found in all strings.
